requirements.txt
speedtest-cli==2.0.2
plotly==2.7.0
